Refactor logic into API and add tests
